requirements.txt
pandas
numpy
scipy
requests
matplotlib
plotly
scikit-learn
statsmodels
gspread
orca
ipykernel
flake8
autopep8
nbformat
openpyxl
patsy
kaleido
bs4
lxml
google-cloud
google-cloud-storage
google-cloud-bigquery
oauth2client
fsspec
gcsfs
schedule